phase 1 of issue #330
